Move classes to search results package
Classes moved:
- PatientSearchResults
- PartitionSearchResultsByVisitedFacility
